+++ dsr_agents/src/agents/person_agent.cpp
@@ -45,7 +45,7 @@ personAgent::personAgent(): AgentNode("person_agent"){
 void personAgent::get_params(){
 	// ROS parameters
 	nav2_util::declare_parameter_if_not_declared(this, "ros_topic", 
-		rclcpp::ParameterValue("/object_detection/detections_3d"), 
+		rclcpp::ParameterValue("detections_3d"), 
 		rcl_interfaces::msg::ParameterDescriptor() 
 			.set__description("The ROS topic to subscribe to"));
 	this->get_parameter("ros_topic", ros_topic_);